+++ client/src/components/todos/TableRowHeader.tsx
@@ -4,9 +4,9 @@ import { useAtom } from 'jotai';
 import { toast } from 'react-toastify';
 
 import { PlainTodo } from '@todo/todo.type';
-import { TABLE_MODALS, PRIMARY_COLORS, TODO_STATE_TEXT, IMPORTANCE_ALPHABET } from '@util/Constants';
+import { PRIMARY_COLORS, TODO_STATE_TEXT, IMPORTANCE_ALPHABET } from '@util/Constants';
 import { copyToClipboard, gethhmmFormat, getyyyymmddDateFormat } from '@util/Common';
-import { modalTypeAtom, todoList, editingTodoIdAtom } from '@util/GlobalState';
+import { todoList, editingTodoIdAtom } from '@util/GlobalState';
 
 import Button from '@components/Button';
 import Image from '@components/Image';
@@ -110,13 +110,14 @@ const TableRowHeader = ({
   prevTodoList,
   nextTodoList,
   onClick,
+  setHasEditModal,
 }: {
   todo: PlainTodo;
   prevTodoList: PlainTodo[];
   nextTodoList: PlainTodo[];
   onClick: MouseEventHandler;
+  setHasEditModal: Function;
 }): ReactElement => {
-  const [, setModalType] = useAtom(modalTypeAtom);
   const [todoListAtom, setTodoListAtom] = useAtom(todoList);
   const [, setEditingTodoId] = useAtom(editingTodoIdAtom);
 
@@ -147,37 +148,39 @@ const TableRowHeader = ({
   const tableRowHeaderElemList = createHeaderElementData({ todo, prevTodoList, nextTodoList });
 
   return (
-    <Wrapper onClick={onClick}>
-      <Button context={<Image src={todo.state === 'DONE' ? Checked : Unchecked} />} onClick={checkTodoStateHandler} />
-      {tableRowHeaderElemList.map((headerElem) => {
-        return (
-          <div key={headerElem.type} style={headerElem.style}>
-            {headerElem.value}
-          </div>
-        );
-      })}
-      <div>
-        <Button
-          context={<img src={Update} width="40px" height="40px" alt="update" />}
-          onClick={() => {
-            setEditingTodoId(todo.id);
-            setModalType(TABLE_MODALS.update);
-          }}
-        />
-        <Button
-          context={<img src={Delete} width="40px" height="40px" alt="delete" />}
-          onClick={() => {
-            handleOnDelete(todo.id);
-          }}
-        />
-        <Button
-          context={<img src={Copy} width="40px" height="40px" alt="copy" />}
-          onClick={() => {
-            copyToClipboard(todo.title);
-          }}
-        />
-      </div>
-    </Wrapper>
+    <>
+      <Wrapper onClick={onClick}>
+        <Button context={<Image src={todo.state === 'DONE' ? Checked : Unchecked} />} onClick={checkTodoStateHandler} />
+        {tableRowHeaderElemList.map((headerElem) => {
+          return (
+            <div key={headerElem.type} style={headerElem.style}>
+              {headerElem.value}
+            </div>
+          );
+        })}
+        <div>
+          <Button
+            context={<img src={Update} width="40px" height="40px" alt="update" />}
+            onClick={() => {
+              setEditingTodoId(todo.id);
+              setHasEditModal(true);
+            }}
+          />
+          <Button
+            context={<img src={Delete} width="40px" height="40px" alt="delete" />}
+            onClick={() => {
+              handleOnDelete(todo.id);
+            }}
+          />
+          <Button
+            context={<img src={Copy} width="40px" height="40px" alt="copy" />}
+            onClick={() => {
+              copyToClipboard(todo.title);
+            }}
+          />
+        </div>
+      </Wrapper>
+    </>
   );
 };
